Refactor, cleanup, add some comments

+++ constants.h
@@ -24,17 +24,19 @@
 static const uint8_t MARK_SLOT1 = 0x08U;
 static const uint8_t MARK_SLOT2 = 0x04U;
 static const uint8_t MARK_NONE  = 0x00U;
+
 static const uint8_t NUMBER_OF_SLOTS = 2U;
 static const long long SLOT_TIME = 30000000LL;     // nanosec
 static const long long TIME_PER_SAMPLE = 41667LL;  // nanosec
-static const long long SAMPLES_PER_SLOT = 720LL;
+static const long long SAMPLES_PER_SLOT = 720LL;   // at 24k
+
 static const unsigned int MAX_MMDVM_CHANNELS = 7U;
 static const unsigned int MAX_PFB_CHANNELS = 12U;  // max sample rate 300k
-static const unsigned int PFB_FILTER_DELAY = 48U;  // must reduce on RPI platforms if bursts are consistently late due to CPU load
-static const unsigned int RESAMPLER_FILTER_DELAY = 24U;
-static const unsigned int TOTAL_FILTER_DELAY = ((2U * RESAMPLER_FILTER_DELAY) + (2U * PFB_FILTER_DELAY)) * TIME_PER_SAMPLE;
+static const unsigned int PFB_FILTER_DELAY = 48U;  // must reduce on RPi platforms if bursts are consistently late due to CPU load
+static const unsigned int RESAMPLER_FILTER_DELAY = 24U; // ntaps = 2 * delay + 1
+static const long long TOTAL_FILTER_DELAY = ((2LL * RESAMPLER_FILTER_DELAY) + (2LL * PFB_FILTER_DELAY)) * TIME_PER_SAMPLE;
 
-static const float TX_DAC_SCALING = 0.35f;
+static const float TX_DAC_SCALING = 0.35f;  // aim for 0 dBm divided across channels at 64 dB gain, Lime PA linear zone
 
 // the following entries are specific for a sample rate of 250k (Lime, Pluto)
 // TODO: handle SXCeiver sample rates
